+++ examples/HRDNS/Socket/EchoServer.php
@@ -120,7 +120,7 @@ class EchoServer extends Command
 
     protected function configure()
     {
-        $this->setName('exmaple:' . strtolower(str_replace('.php', '', basename(__FILE__))));
+        $this->setName('example:' . strtolower(str_replace('.php', '', basename(__FILE__))));
         $this->setProcessTitle($this->getName());
         $this->setDescription('A TCPServer EchoServer');
         $this->setHelp('');