+++ examples/HRDNS/Socket/WebSocketServer.php
@@ -128,7 +128,7 @@ class WebSocketServer extends Command
 
     protected function configure()
     {
-        $this->setName('exmaple:' . strtolower(str_replace('.php', '', basename(__FILE__))));
+        $this->setName('example:' . strtolower(str_replace('.php', '', basename(__FILE__))));
         $this->setProcessTitle($this->getName());
         $this->setDescription('A WebSocket Echo Server');
         $this->setHelp('');