+++ src/HRDNS/Socket/Client/TCPClient.php
@@ -5,11 +5,25 @@ namespace HRDNS\Socket\Client;
 class TCPClient extends Client
 {
 
+    /** @var bool */
+    private $blocking = false;
+
+    /**
+     * @param bool $blocking
+     * @return TCPClient
+     */
+    public function setBlockingMode(bool $blocking)
+    {
+        $this->blocking = $blocking;
+
+        return $this;
+    }
+
     /**
      * @return self
      * @throws \Exception
      */
-    public function connect(): self
+    public function connect()
     {
         if (is_resource($this->socket)) {
             return $this;
@@ -33,19 +47,19 @@ class TCPClient extends Client
             $this->socket,
             SOL_SOCKET,
             SO_RCVTIMEO,
-            array(
+            [
                 'sec' => $this->timeoutSeconds,
-                'usec' => $this->timeoutUSeconds
-            )
+                'usec' => $this->timeoutUSeconds,
+            ]
         );
         @socket_set_option(
             $this->socket,
             SOL_SOCKET,
             SO_SNDTIMEO,
-            array(
+            [
                 'sec' => $this->timeoutSeconds,
-                'usec' => $this->timeoutUSeconds
-            )
+                'usec' => $this->timeoutUSeconds,
+            ]
         );
 
         if (@socket_connect($this->socket, $this->host, $this->port) === false) {
@@ -61,27 +75,30 @@ class TCPClient extends Client
                 )
             );
         }
-        socket_set_nonblock($this->socket);
+        if (!$this->blocking) {
+            socket_set_nonblock($this->socket);
+        }
 
         return $this;
     }
 
     /**
-     * @todo fix mixed return types!
-     * @param integer|null $length
-     * @return string|boolean
+     * @param int|null $length
+     * @param mixed|null $type
+     * @return boolean|string
      */
-    public function read(int $length = null)
+    public function read(int $length = null, $type = null)
     {
+        $type = $type === null ? PHP_NORMAL_READ : $type;
         $length = $length === null ? $this->bufferLength : $length;
         if (!is_resource($this->socket) || !$length) {
             return false;
         }
-        return @socket_read($this->socket, $length);
+
+        return @socket_read($this->socket, $length, $type);
     }
 
     /**
-     * @todo fix mixed return types!
      * @param string $buffer
      * @param integer|null $length
      * @return integer|boolean
@@ -92,6 +109,7 @@ class TCPClient extends Client
             return false;
         }
         $length = $length === null ? strlen($buffer) : $length;
+
         return @socket_write($this->socket, $buffer, $length);
     }